Change type of offset from int to int16_t
This ensures that when run on 32-bit platforms that have longer ints
(32-bits), that the storage in EEPROM is not wasteful of the extra 2
bytes.

+++ Timezone.h
@@ -32,7 +32,7 @@ struct TimeChangeRule
     uint8_t dow;       //day of week, 1=Sun, 2=Mon, ... 7=Sat
     uint8_t month;     //1=Jan, 2=Feb, ... 12=Dec
     uint8_t hour;      //0-23
-    int offset;        //offset from UTC in minutes
+    int16_t offset;    //offset from UTC in minutes
 };
         
 class Timezone
